refactor: move platform specific code to platform specific modules
