+++ src/tracing/net/ipv6.rs
@@ -1,6 +1,7 @@
 use crate::tracing::error::TracerError::AddressNotAvailable;
 use crate::tracing::error::{TraceResult, TracerError};
 use crate::tracing::net::channel::MAX_PACKET_SIZE;
+use crate::tracing::net::platform;
 use crate::tracing::packet::checksum::{icmp_ipv6_checksum, udp_ipv6_checksum};
 use crate::tracing::packet::icmpv6::destination_unreachable::DestinationUnreachablePacket;
 use crate::tracing::packet::icmpv6::echo_reply::EchoReplyPacket;
@@ -14,8 +15,7 @@ use crate::tracing::probe::{ProbeResponse, ProbeResponseData, TcpProbeResponseDa
 use crate::tracing::types::{PacketSize, PayloadPattern, Sequence, TraceId};
 use crate::tracing::util::Required;
 use crate::tracing::{PortDirection, Probe, TracerProtocol};
-use nix::sys::socket::{AddressFamily, SockaddrLike};
-use socket2::{Domain, Protocol, SockAddr, Socket, Type};
+use socket2::{SockAddr, Socket};
 use std::io::ErrorKind;
 use std::net::{IpAddr, Ipv6Addr, Shutdown, SocketAddr};
 use std::time::SystemTime;
@@ -31,39 +31,6 @@ const MAX_ICMP_PACKET_BUF: usize = MAX_PACKET_SIZE - Ipv6Packet::minimum_packet_
 
 /// The maximum size of ICMP payload we allow.
 const MAX_ICMP_PAYLOAD_BUF: usize = MAX_ICMP_PACKET_BUF - IcmpPacket::minimum_packet_size();
-
-pub fn lookup_interface_addr(name: &str) -> TraceResult<IpAddr> {
-    nix::ifaddrs::getifaddrs()
-        .map_err(|_| TracerError::UnknownInterface(name.to_string()))?
-        .into_iter()
-        .find_map(|ia| {
-            ia.address.and_then(|addr| match addr.family() {
-                Some(AddressFamily::Inet6) if ia.interface_name == name => addr
-                    .as_sockaddr_in6()
-                    .map(|sock_addr| IpAddr::V6(sock_addr.ip())),
-                _ => None,
-            })
-        })
-        .ok_or_else(|| TracerError::UnknownInterface(name.to_string()))
-}
-
-pub fn make_icmp_send_socket() -> TraceResult<Socket> {
-    let socket = Socket::new(Domain::IPV6, Type::RAW, Some(Protocol::ICMPV6))?;
-    socket.set_nonblocking(true)?;
-    Ok(socket)
-}
-
-pub fn make_udp_send_socket() -> TraceResult<Socket> {
-    let socket = Socket::new(Domain::IPV6, Type::RAW, Some(Protocol::UDP))?;
-    socket.set_nonblocking(true)?;
-    Ok(socket)
-}
-
-pub fn make_recv_socket() -> TraceResult<Socket> {
-    let socket = Socket::new(Domain::IPV6, Type::RAW, Some(Protocol::ICMPV6))?;
-    socket.set_nonblocking(true)?;
-    Ok(socket)
-}
 
 pub fn dispatch_icmp_probe(
     icmp_send_socket: &mut Socket,
@@ -147,9 +114,7 @@ pub fn dispatch_tcp_probe(
         PortDirection::FixedDest(dest_port) => (probe.sequence.0, dest_port.0),
         PortDirection::FixedBoth(_, _) | PortDirection::None => unimplemented!(),
     };
-    let socket = Socket::new(Domain::IPV6, Type::STREAM, Some(Protocol::TCP))?;
-    socket.set_nonblocking(true)?;
-    socket.set_reuse_port(true)?;
+    let socket = platform::make_stream_socket_ipv6()?;
     let local_addr = SocketAddr::new(IpAddr::V6(src_addr), src_port);
     socket.bind(&SockAddr::from(local_addr))?;
     socket.set_unicast_hops_v6(u32::from(probe.ttl.0))?;
@@ -158,7 +123,7 @@ pub fn dispatch_tcp_probe(
         Ok(_) => {}
         Err(err) => {
             if let Some(code) = err.raw_os_error() {
-                if nix::Error::from_i32(code) != nix::Error::EINPROGRESS {
+                if platform::is_in_progress_error(code) {
                     return match err.kind() {
                         ErrorKind::AddrInUse | ErrorKind::AddrNotAvailable => {
                             Err(AddressNotAvailable(local_addr))
@@ -210,7 +175,7 @@ pub fn recv_tcp_socket(
         }
         Some(err) => {
             if let Some(code) = err.raw_os_error() {
-                if nix::Error::from_i32(code) == nix::Error::ECONNREFUSED {
+                if platform::is_conn_refused_error(code) {
                     return Ok(Some(ProbeResponse::TcpRefused(TcpProbeResponseData::new(
                         SystemTime::now(),
                         dest_addr,
