+++ src/tracing/net/platform/windows.rs
@@ -1,0 +1,81 @@
+use super::byte_order::PlatformIpv4FieldByteOrder;
+use crate::tracing::error::TraceResult;
+use socket2::Socket;
+use std::net::IpAddr;
+use std::time::Duration;
+
+/// TODO
+#[allow(clippy::unnecessary_wraps)]
+pub fn for_address(_src_addr: IpAddr) -> TraceResult<PlatformIpv4FieldByteOrder> {
+    Ok(PlatformIpv4FieldByteOrder::Network)
+}
+
+/// TODO
+pub fn lookup_interface_addr_ipv4(_name: &str) -> TraceResult<IpAddr> {
+    unimplemented!()
+}
+
+/// TODO
+pub fn lookup_interface_addr_ipv6(_name: &str) -> TraceResult<IpAddr> {
+    unimplemented!()
+}
+
+/// TODO
+pub fn make_icmp_send_socket_ipv4() -> TraceResult<Socket> {
+    unimplemented!()
+}
+
+/// TODO
+pub fn make_udp_send_socket_ipv4() -> TraceResult<Socket> {
+    unimplemented!()
+}
+
+/// TODO
+pub fn make_recv_socket_ipv4() -> TraceResult<Socket> {
+    unimplemented!()
+}
+
+/// TODO
+pub fn make_icmp_send_socket_ipv6() -> TraceResult<Socket> {
+    unimplemented!()
+}
+
+/// TODO
+pub fn make_udp_send_socket_ipv6() -> TraceResult<Socket> {
+    unimplemented!()
+}
+
+/// TODO
+pub fn make_recv_socket_ipv6() -> TraceResult<Socket> {
+    unimplemented!()
+}
+
+/// TODO
+pub fn make_stream_socket_ipv4() -> TraceResult<Socket> {
+    unimplemented!()
+}
+
+/// TODO
+pub fn make_stream_socket_ipv6() -> TraceResult<Socket> {
+    unimplemented!()
+}
+
+/// TODO
+pub fn is_readable(_sock: &Socket, _timeout: Duration) -> TraceResult<bool> {
+    unimplemented!()
+}
+
+/// TODO
+pub fn is_writable(_sock: &Socket) -> TraceResult<bool> {
+    unimplemented!()
+}
+
+/// TODO
+pub fn is_in_progress_error(_code: i32) -> bool {
+    unimplemented!()
+}
+
+/// TODO
+pub fn is_conn_refused_error(_code: i32) -> bool {
+    unimplemented!()
+}
